namespaces & symfony tests
